fix: se aplica responsabilidad unica (SOLID) al codigo
