+++ domain/usecase/src/main/java/co/com/pragma/usecase/requests/interfaces/SqsUseCaseInterface.java
@@ -1,4 +1,4 @@
-package co.com.pragma.usecase.requests;
+package co.com.pragma.usecase.requests.interfaces;
 
 import co.com.pragma.usecase.requests.dto.SqsMessageDTO;
 import reactor.core.publisher.Mono;
