+++ domain/usecase/src/main/java/co/com/pragma/usecase/requests/interfaces/UserUseCaseInterface.java
@@ -1,4 +1,4 @@
-package co.com.pragma.usecase.requests;
+package co.com.pragma.usecase.requests.interfaces;
 
 import co.com.pragma.usecase.requests.dto.UserSalaryInformationDTO;
 import reactor.core.publisher.Mono;
