+++ domain/usecase/src/main/java/co/com/pragma/usecase/requests/validations/cases/UserExistValidation.java
@@ -1,7 +1,7 @@
 package co.com.pragma.usecase.requests.validations.cases;
 
 import co.com.pragma.model.requests.Requests;
-import co.com.pragma.usecase.requests.UserUseCaseInterface;
+import co.com.pragma.usecase.requests.interfaces.UserUseCaseInterface;
 import co.com.pragma.usecase.requests.validations.IRequestsValidation;
 import co.com.pragma.usecase.requests.validations.error.RequestsValidationException;
 import reactor.core.publisher.Mono;
